Make exception subcode a long
On EXC_BAD_ACCESS, exception subcode is used to pass the faulting memory
address, so it needs to be (at least) pointer-sized. Thus, make it into
a long.

This requires matching changes in glibc and the Hurd.
Message-Id: <20230319151017.531737-4-bugaevc@gmail.com>

+++ i386/i386/trap.h
@@ -40,7 +40,7 @@ void
 i386_exception(
 	int	exc,
 	int	code,
-	int	subcode) __attribute__ ((noreturn));
+	long	subcode) __attribute__ ((noreturn));
 
 extern void
 thread_kdb_return(void);
